move alert interval creation into alert invariant
