+++ pkg/invariants/alertanalyzer/alerts_test.go
@@ -1,4 +1,4 @@
-package monitor
+package alertanalyzer
 
 import (
 	"reflect"
